requirements.txt
streamlit~=1.39.0
streamlit-option-menu
streamlit_card
pillow~=10.4.0
google.generativeai
protobuf~=5.28.2
requests
streamlit-card
PyPDF2
docx
exceptions